+++ docker-compose.yml
@@ -1,0 +1,15 @@
+version: "3.9"
+
+services:
+  aruco_app:
+    build: .
+    container_name: aruco_pose_container
+    volumes:
+      - ./src/pose_estimation.py:/app/pose_estimation.py
+      - ./src/input:/app/input
+      - ./src/output:/app/output
+      - ./src/log:/app/log
+    working_dir: /app
+    command: bash
+    stdin_open: true
+    tty: true
