+++ src/requirements.txt
@@ -1,0 +1,4 @@
+opencv-python==4.10.0.84
+opencv-contrib-python==4.10.0.84
+numpy==1.26.2
+matplotlib==3.8.2
